+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/contentful/contentful-core.php/compare/2.1.1...HEAD)
 
+<!-- PENDING-CHANGES -->
+> No meaningful changes since last release.
+<!-- /PENDING-CHANGES -->
+
 ## [2.1.1](https://github.com/contentful/contentful-core.php/tree/2.1.1) (2018-11-08)
 
 ### Fixed